basemaps: nasa: add datepicker
