+++ app/assets/stylesheets/common/datepicker.css.scss
@@ -94,6 +94,41 @@ $sDates-defaultWidth: 434px;
   border-top: 10px solid #CCC;
   z-index: 0;
 }
+.DatePicker-simpleDropdown {
+  position: absolute;
+  right: -16px;
+  top: 49px;
+  width: 236px;
+  padding: 0;
+  border: 1px solid #CCC;
+  font: normal 13px/20px "Lato",sans-serif;
+  background: white;
+  border-radius: $border-radius;
+  box-shadow: rgba(black, 0.1) 0 0 2px 2px;
+}
+.DatePicker-simpleDropdown:after,
+.DatePicker-simpleDropdown:before {
+  content: '';
+  position: absolute;
+  left: 50%;
+  top: -10px;
+}
+.DatePicker-simpleDropdown:after {
+  top: -8px;
+  margin-left: -4px;
+  border-left: 8px solid transparent;
+  border-right: 8px solid transparent;
+  border-bottom: 8px solid $cCard-hoverFill;
+  z-index: 1;
+}
+.DatePicker-simpleDropdown:before {
+  top: -10px;
+  margin-left: -6px;
+  border-left: 10px solid transparent;
+  border-right: 10px solid transparent;
+  border-bottom: 10px solid #CCC;
+  z-index: 0;
+}
 
 // Timers
 .DatePicker-timers {
@@ -254,6 +289,7 @@ td.datepickerSpecial a {
 td.datepickerSpecial.datepickerSelected a {
   background: #a00;
 }
+
 
 //////////////////////////////
 // Custom datepicker styles //
