--- lib/assets/javascripts/cartodb3/components/date-picker/calendar-dropdown.tpl
@@ -1,3 +1,0 @@
-<div class="DatePicker">
-  <div class="js-calendar DatePicker-calendar DatePicker-calendar--simple"></div>
-</div>
