--- lib/assets/javascripts/cartodb3/components/date-picker/date-picker-form.tpl
@@ -1,22 +1,0 @@
-<form>
-  <div class="DatePicker-timersFrom">
-    <div class="DatePicker-timersHour">
-      <label class="DatePicker-timersLabel u-upperCase"><%- _t('components.datepicker.hour') %></label>
-      <div data-editors="fromHour"></div>
-    </div>
-    <div class="DatePicker-timersMin">
-      <label class="DatePicker-timersLabel u-upperCase"><%- _t('components.datepicker.min') %></label>
-      <div data-editors="fromMin"></div>
-    </div>
-  </div>
-  <div class="DatePicker-timersTo">
-    <div class="DatePicker-timersHour">
-      <label class="DatePicker-timersLabel u-upperCase"><%- _t('components.datepicker.hour') %></label>
-      <div data-editors="toHour"></div>
-    </div>
-    <div class="DatePicker-timersMin">
-      <label class="DatePicker-timersLabel u-upperCase"><%- _t('components.datepicker.min') %></label>
-      <div data-editors="toMin"></div>
-    </div>
-  </div>
-</form>
