--- lib/assets/javascripts/cartodb3/components/date-picker/date-picker-range-form.tpl
@@ -1,14 +1,0 @@
-<form class="DatePicker-timers">
-  <div class="DatePicker-timersField u-rSpace--xl">
-    <div class="DatePicker-timersInput" data-fields="fromHour"></div>
-  </div>
-  <div class="DatePicker-timersField u-rSpace--xl">
-    <div class="DatePicker-timersInput" data-fields="fromMin"></div>
-  </div>
-  <div class="DatePicker-timersField DatePicker-timersField--spaced u-rSpace--xl">
-    <div class="DatePicker-timersInput" data-fields="toHour"></div>
-  </div>
-  <div class="DatePicker-timersField">
-    <div class="DatePicker-timersInput" data-fields="toMin"></div>
-  </div>
-</form>
