+++ lib/assets/javascripts/cartodb3/components/date-picker/date-picker.tpl
@@ -1,0 +1,7 @@
+<button class="js-dates DatePicker-dates DatePicker-dates--singleDate js-DatePicker-dates has-icon <% if (readOnly) { %> is-disabled <% } %>">
+  <strong class="js-date-str"><%- date %></strong>
+  <i class="CDB-IconFont CDB-IconFont-calendar DatePicker-datesIcon"></i>
+</button>
+<div class="DatePicker-simpleDropdown js-DatePicker-simpleDropdown">
+  <div class="js-calendar DatePicker-calendar DatePicker-calendar--simple"></div>
+</div>
