+++ lib/assets/javascripts/cartodb3/components/date-picker/edit-field-model.js
