--- lib/assets/javascripts/cartodb3/components/modals/add-basemap/nasa/edit-fields/edit-field.tpl
@@ -1,1 +1,0 @@
-<input type="text" class="Form-input" value="<%- value %>" <% if (readOnly) { %> readonly <% } %>/>
