+++ lib/assets/javascripts/cartodb3/components/modals/add-basemap/nasa/nasa-view.js
@@ -1,9 +1,10 @@
+var $ = require('jquery');
 var moment = require('moment');
 var CoreView = require('backbone/core-view');
-var DateFieldView = require('../../../edit-fields/date-field/date-field-view');
-var EditFieldModel = require('./edit-fields/edit-field-model');
+var DatePickerView = require('../../../../components/date-picker/date-picker-view');
+var EditFieldModel = require('../../../../components/date-picker/edit-field-model');
 var TipsyTooltipView = require('../../../../components/tipsy-tooltip-view');
-var template = require('./enter-url.tpl');
+var template = require('./nasa.tpl');
 
 /**
  * Represents the NASA tab content
@@ -40,10 +41,11 @@ module.exports = CoreView.extend({
 
     this.$el.html(
       template({
-        layerType: this.model.get('layerType'),
-        initialDateStr: moment(this.dateModel.get('value')).format(this.options.dateFormat)
+        layerType: this.model.get('layerType')
       })
     );
+
+    this._renderDatePicker();
 
     return this;
   },
@@ -54,22 +56,16 @@ module.exports = CoreView.extend({
 
       this.render();
     }, this);
-    this.dateModel.bind('change:value', function () {
-      var date = moment(this.dateModel.get('value')).format(this.options.dateFormat);
-      this.model.set('date', date);
-    }, this);
-    this.add_related_model(this.dateModel);
   },
 
   _renderDatePicker: function () {
-    // Date field
-    var dateField = new DateFieldView({
-      model: this.dateModel,
-      showTime: false,
-      showGMT: false
+    // Date picker
+    var datepicker = this.datepicker = new DatePickerView({
+      className: 'DatePicker DatePicker--withBorder',
+      model: this.dateModel
     });
-    this.addView(dateField);
-    this._$datePicker().append(dateField.render().el);
+    this._$datePicker().append(datepicker.render().el);
+    this.addView(datepicker);
 
     // Disabled tooltip
     if (this.dateModel.get('readOnly')) {
