+++ lib/assets/javascripts/cartodb3/components/modals/add-basemap/tilejson/tilejson-view.js
@@ -15,11 +15,9 @@ module.exports = CoreView.extend({
 
   initialize: function (opts) {
     if (!opts.submitButton) throw new Error('submitButton is required');
-    if (!opts.modalFooter) throw new Error('modalFooter is required');
     if (!opts.tileJSONLayerModel) throw new Error('tileJSONLayerModel is required');
 
     this._submitButton = opts.submitButton;
-    this._modalFooter = opts.modalFooter;
     this._tileJSONLayerModel = opts.tileJSONLayerModel;
     this._lastURL = '';
     this._debouncedUpdate = _.debounce(this._update.bind(this), 150);
