+++ lib/assets/javascripts/cartodb3/components/modals/add-layer/content/imports/import-twitter/import-twitter-view.js
@@ -2,7 +2,7 @@ var _ = require('underscore');
 var moment = require('moment');
 var ImportView = require('../import-view');
 var UploadModel = require('../../../../../../data/upload-model');
-var DatesRangePickerView = require('../../../../../date-picker/date-picker-range-view');
+var DatesRangePickerView = require('../../../../../date-picker-range/date-picker-range-view');
 var TwitterCategoriesView = require('./twitter-categories/twitter-categories-view');
 var CreditsUsageView = require('./credits-usage-view.js');
 var template = require('./import-twitter.tpl');
